replace typeconv in memlog
